// web/src/shared/api/generated/model/successRateMetricsResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { SuccessRateGroupResult } from './successRateGroupResult';

export interface SuccessRateMetricsResult {
  from: string;
  to: string;
  /** @nullable */
  subjectId: string | null;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  attempts: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  correct: number | string;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
     */
  rate: number | string | null;
  bySubject: SuccessRateGroupResult[];
  byUnit: SuccessRateGroupResult[];
  byLesson: SuccessRateGroupResult[];
  generatedAt: string;
}